refactor: make real node12+ ESM module
